Add component for market detail
